+++ src/Project/AbstractProject.php
@@ -15,6 +15,7 @@ namespace Mds\PimPrint\CoreBundle\Project;
 
 use Mds\PimPrint\CoreBundle\InDesign\Command\AbstractCommand;
 use Mds\PimPrint\CoreBundle\InDesign\CommandQueue;
+use Mds\PimPrint\CoreBundle\Project\Traits\BoxIdentTrait;
 use Mds\PimPrint\CoreBundle\Project\Traits\FormFieldsTrait;
 use Mds\PimPrint\CoreBundle\Project\Traits\ServicesTrait;
 use Mds\PimPrint\CoreBundle\Project\Traits\RenderingTrait;
@@ -43,6 +44,7 @@ abstract class AbstractProject
     use TemplateTrait;
     use RenderingTrait;
     use FormFieldsTrait;
+    use BoxIdentTrait;
 
     /**
      * CommandQueue instance.
@@ -189,7 +191,7 @@ abstract class AbstractProject
      * @return string
      * @throws \Exception
      */
-    protected function getLanguage()
+    public function getLanguage()
     {
         return $this->pluginParams()
                     ->get(PluginParameters::PARAM_LANGUAGE);